requirements.txt
flask
transformers
torch
google-cloud-texttospeech
python-dotenv